+++ src/pages/Trading/components/TradingModal/LocationModal.tsx
