+++ src/pages/Trading/components/TradingModal/MoneyModal.tsx
@@ -1,27 +1,32 @@
-import React from 'react';
+import React, { useState, useRef } from 'react';
 import styled from 'styled-components';
-import { BREAK_POINT } from '../../../constants/style';
-import Text from '../../../components/Text';
+import { BREAK_POINT } from '../../../../constants/style';
+import RangeSlider from '../../../../components/RangeSlider';
 
-interface FilterModalProps {
+interface MoneyModalProps {
   isOpen: boolean;
 }
-
-const FilterModal = ({ isOpen }: FilterModalProps) => {
+const MoneyModal = ({ isOpen }: MoneyModalProps) => {
   return (
     <Dimmed isOpen={isOpen}>
-      <Container onClick={e => e.stopPropagation()}>
-        <TextWrapper style={{ borderBottom: '1px solid #E5E5E5' }}>{'찜이 많은 순'}</TextWrapper>
-        <TextWrapper>{'추천 순'}</TextWrapper>
+      <Container
+        onClick={e => {
+          e.stopPropagation();
+        }}
+      >
+        <span style={{ marginTop: '16px', marginLeft: '16px' }}>{'가격'}</span>
+        <SliderBox>
+          <RangeSlider />
+        </SliderBox>
       </Container>
     </Dimmed>
   );
 };
 
-export default FilterModal;
-
+export default MoneyModal;
 const Dimmed = styled.div<{ isOpen: boolean }>`
   visibility: ${props => (props.isOpen ? 'visible' : 'hidden')};
+  position: relative;
 
   @media screen and (max-width: ${BREAK_POINT.MOBILE}) {
     position: fixed;
@@ -40,11 +45,11 @@ const Dimmed = styled.div<{ isOpen: boolean }>`
 const Container = styled.div`
   position: absolute;
   width: 306px;
-  height: 60px;
+  height: 162px;
   background-color: #ffffff;
   z-index: 1000;
-  left: 0px;
-  top: 37px;
+  left: -70px;
+  top: 20px;
   display: flex;
   flex-direction: column;
   border-radius: 10px;
@@ -58,10 +63,9 @@ const Container = styled.div`
   }
 `;
 
-const TextWrapper = styled.div`
-  line-height: 24px;
-  margin-top: 5px;
-  margin-left: 16px;
+const SliderBox = styled.div`
   width: 100%;
-  height: 50%;
+  height: 100%;
+  margin-left: 26px;
+  margin-top: 84px;
 `;
